Extending "internal_dirs" to be exensible and configurable.
Simply add a 'spine_local_internals_dirs' key in /spine_internals/config/
with a list of directories to process *as-if* they were part of spine_internals.

This list should *not* include the "config" part, but the path *should* have a "config" directory under it (though spine will ignore it if it doesn't).

In other words, if you want to include:

  /mycoolgroup/spine_internals/config/some_piece_of_early_data

then just do

  mycoolgroup/spine_internals

inside of spine_local_internals_dirs and spine will do the rest for you. Remember, like the rest of spine-internals you will NOT have the rest of your keys defined, this is special early boot-strapping stuff only. Abusing this will cause the bullet to go squarely into your foot.

- Phil