From: Chet Burgess <chester.burgess@ticketmaster.com>
The use of stat's rdev on 2.4 kernels is unreliable. Here we only use it on
non-2.4 kernels. See issue #15.

An earlier version of this patch was committed and reverted, and this is a
follow up with additional considerations requested by Ryan Tilder.

This one looks good and has been tested, so I'm committing.

Signed-off-by: Chet Burgess <chester.burgess@ticketmaster.com>
Signed-off-by: Phil Dibowitz <phil@ipom.com>