From: Chet Burgess <chester.burgess@ticketmaster.com>
OK so ultimately the Hardware plugin should be able to provide this information.

That being said I took a look at the current state of the hardware plugin and it is
ugly. The plugin won't even run complaining about numerous syntax errors and
undefined variables. It will probably take a good amount of effort to get it working
at this point.

So as a temporary hack (since we are in need of a solution to this) I have added
support to the SystemInfo plugin to use dmidecode to return the "Product Name" string
in the c_hardware_platform key. If for some reason dmidecode does not return a string
(like it will on Xen VMs) the key will be set to "UNKNOWN".

Signed-off-by: Phil Dibowitz <phil@ipom.com>