From: Chet Burgess <chester.burgess@ticketmaster.com>
I have re-written the is_virtual function in the SystemInfo plug-in to properly
detect VMware, Xen Para and Xen HVM. This necessity breaking backwards compatibility
to implement cleanly. As TM is the only existing user we have internally agreed to
deal with changing our tree to function with the new code. 

Additionally the new is_virtual plug-in no longer automagically includes a role for
virtual support as we now have multiple types of virtual systems and the inclusion of
roles is something that should be done in the data tree not in the code.

 * Also, various style cleanups by Phil Dibowitz <phil@ipom.com>

Signed-off-by: Phil Dibowitz <phil@ipom.com>