From: Chet Burgess <chester.burgess@ticketmaster.com>
While my patch works great it does break backwards compatibility with with
existing data trees. I had hoped that it would be possible to adjust existing
trees to account for this. 

I have however since discovered that this will be difficult for certain older,
pre spine, version of the code (specially the 1.6.6 and earlier branches of
code).

As such I am resubmitting a patch that restores the old functionality to
c_is_virtual. The c_is_virtual key should be considered deprecated and as soon
as all older trees are upgrades (of which there are exactly 2) I will submit a
new patch to remove the c_is_virtual entirely and stick purely with the
c_virtual_type variable.

Issue 26 opened to undo horrible hack in the near future.

Additionally, style cleanups through out the file from Phil Dibowitz
<phil@ipom.com>.

Signed-off-by: Phil Dibowitz <phil@ipom.com>