From: Charlie Brown <charlie.brown@ticketmaster.com>
This patch addresses issue 40. It addes support to Auth.pm for per user
and group home directory permissions as well as ability to specifiy default
permissions for home directories.

Charlie Brown is responsible for the general idea, as well as the code for
supporting the "permissions" element in an auth object. Chet Burgess did some
basic cleanup and the coding for support of the default home directory
permissions.

--
Chet Burgess

Signed-off-by: Phil Dibowitz <phil@ipom.com>
Signed-off-by: Chet Burgess <cfb@liquidreality.org>