From: Chet Burgess <cfb@liquidreality.org>
Commiting changes to fix issue 38.

Overlay now reads the name of directories to be parsed as overlays and 
their targets from a key called "overlay_map". If the "overlay_map" key is
empty, it assumes the only directory to parse is called "overlay" and that
its target on the filesystem is /.

The "overlay_map" key has the form of "directory:target" where directory is a
directory in the config source and target is the absolute path on the file
system to place the contents of the directories.

Overlay then iterates through all elements of the "overlay_map" key and
processes each one in turn.

Signed-off-by: Chet Burgess <cfb@liquidreality.org>