Commiting fixes for issue 42 and 43.
Issue 42:
Adding support for a function called get_plugin_version which can be 
queried from the c object in keys and templates and will return 
the version of a given plugin.

Issue 43:
Fixing automatic version strings in the various perl modules
to use the correct form for Subversion check-in numbers.

Chet Burgess <cfb@liquidreality.org>