New functionality for issue 44.
The attached patch removes the list of hardcoded files to copy and now
looks for a key called firstboot_cpfiles. The firstboot_cpfiles key as the
form "file,directory" where file is the full path to the "file" to be
copied and "directory" is the full path to the directory to copy "file"
into. If "directory" does not exist, FirstBoot will create it with its
default mode (value of "firstrun_default_mode" key or 0755 if not found)
and default uid and gid (value of firstrun_default_ugid" or 0:0 if not found).

--
Chet Burgess <cfb@liquidreality.org>