Fixes for issue #13 and issue #37.
issue #13:
Pass the package name to the cleanup function as 
"package" instead of "package.32bit" as the cleanup 
code does not support architecture specific packages.

issue #37:
Pull the value of the "rpm_opts" key from the config 
source and pass it on the command line to the rpm -e
command when removing packages (add --allmatches to
rpm_opts to remove all versions of the package with
the same name).