Fix for issue #89.
SystemInfo will now look for a key called network_path that provides the path
to where the the network directory is located. If the key is not present the
code will default to simply "network" (the current behavior). 

The network_path key will need to be placed in the spine_internal directory as
SystemInfo runs during the DISCOVERY/populate phase before the tree is parsed.