Fix for issue #80.
Use lstat instead of stat in order to ensure we are testing the permissions
of the symlink and not the underlying file. The underlying file should be 
caught about the search anyway. 

Additionally log if we encounter a symlink with the SUID and/or SGID bit set
as we can't reliably change the permissions of the symlink.